RB-476 - Take a branch and rename the classes
