+++ kafka-connect-aws-s3/src/it/scala/io/lenses/streamreactor/connect/aws/s3/source/S3SourceAvroEnvelopeTest.scala
@@ -122,7 +122,7 @@ class S3SourceAvroEnvelopeTest extends S3ProxyContainerTest with AnyFlatSpecLike
 
   "task" should "extract from avro files containing the envelope" in {
 
-    val task = new S3SourceTask()
+    val task = new S3SourceTaskDeprecated()
 
     val props = DefaultProps
       .combine(
